requirements.txt
boto3
numpy
protobuf
python-dotenv
Requests
sagemaker
streamlit
streamlit_option_menu